pyproject.toml
[tool.poetry]
name = "firmware"
version = "1.0.0"
description = "Configuration files for AVR systems connected to intelligent relays"
authors = ["Martynov Viktor"]
license = "Eclipse Public License - v 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
pillow = "10.2.0"
django-imagekit = "5.0.0"
python-dotenv = "1.0.0"
django-filter = "23.5"
markupsafe = "2.1.3"
click = "8.1.7"
django-mptt = "^0.16.0"
django-mptt-admin = "^2.6.1"
django-import-export = "^3.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
